+++ zarrs_storage/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- `MemoryStore`: avoid unnecessary value cloning in `get` and `get_partial_many`
+
 ## [0.4.2] - 2026-01-13
 
 ### Added